requirements.txt
easyocr==1.7.2
filelock==3.18.0
fsspec==2025.7.0
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
pyclipper==1.3.0.post6
pytesseract==0.3.13
python-bidi==0.6.6
PyYAML==6.0.2
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.1
sympy==1.14.0
tifffile==2025.5.10
torch==2.7.1
torchvision==0.22.1
typing_extensions==4.14.1
